building tests

+++ sql_scripts/recordings.sql
@@ -5,7 +5,7 @@ CREATE TABLE recordings(
     week_day INT[], -- Filled with values of 0-6 to denote what days of the week to record on
     start_date DATE NOT NULL,
     stop_date DATE,
-    duration FLOAT(4, 2) NOT NULL,
+    duration FLOAT(4) NOT NULL,
     start_time TIME NOT NULL,
     last_started TIMESTAMP,
     camera_topics TEXT[] NOT NULL
